Drop devise on pages and install
Devise is a can of worms, and not all devs want to have Rails UI dictate auth strategies. It feels to heavy handed to assume this.

Instead of integrating devise directly with pages I’ll proceed forward to provide snippets in the component library instead. This simplifies a great deal and offers more devs possiblities to use Rails UI.

More work is necessary to provide status snippets and devise specific snippets but it’s not far off.